- code re-shuffle - removed r3d_animtree module, all related structs and functions moved to r3d_animation_tree - added common animation utility functions file - updated r3d_animation_player (not fully done yet)
