+++ CMakeLists.txt
@@ -251,12 +251,12 @@ add_library(${PROJECT_NAME}
     # Importer
     ${R3D_IMPORTER_SOURCES}
     # Common
+    "${R3D_ROOT_PATH}/src/common/r3d_anim.c"
     "${R3D_ROOT_PATH}/src/common/r3d_frustum.c"
     "${R3D_ROOT_PATH}/src/common/r3d_helper.c"
     "${R3D_ROOT_PATH}/src/common/r3d_image.c"
     "${R3D_ROOT_PATH}/src/common/r3d_pass.c"
     # Modules
-    "${R3D_ROOT_PATH}/src/modules/r3d_animtree.c"
     "${R3D_ROOT_PATH}/src/modules/r3d_texture.c"
     "${R3D_ROOT_PATH}/src/modules/r3d_target.c"
     "${R3D_ROOT_PATH}/src/modules/r3d_shader.c"
